requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
idna==3.10
pycparser==2.22
pyodbc==5.2.0
python-dotenv==1.1.0
requests==2.32.4
urllib3==2.4.0
